Initial commit of MinnowBoard-Max Extras
This initial commit contributes a variety of tests for the calamari and
hse MinnowBoard-Max Lures. The calamari.c boardfile is meant as a rapid
prototyping and testing aid, and not for the upstream Linux kernel.

Future examples will be added demonstrating the use of ACPI _DSD Device
Properties for the description of Lures, as well as additional Lure
tests and prototyping files.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>