LSE: Add Linux GPIO discovery tool what_gpio_num.sh
This adds lse/what_gpio_num.sh to allow for easier discovery of what the GPIO
numbers under Linux *SHOULD* be.  This is coming out of the fact that the
wiki doesn't seem to be accurate for all platforms (specifically Ubuntu
seems to differ and it's not clear why yet).

There are two variables that can be set to modify the behavior:

DEBUG set to non-zero will output some extra debug data

SINGLECOL set to non-zero will output the table in a single column

Signed-off-by: John 'Warthog9' Hawley <john.hawley@intel.com>