Adding offset test for GPIOs for kernels 3.18 and later
In kernel 3.18 the enumeration of GPIOs changed.  They count down
from 256 in 3.17 and earlier, and down from 512 in 3.18 and later.
This means that in newer kernels the GPIOs are offset by 256, and
obviously break for all of the things that were trying to use them.
This adds a kernel version check that should work to auto select
the right GPIOs.